Adding a couple packages to the image
It was suggested to add git to the image.  That seemed reasonable, as it would make it easy to pull in files from GitHub.  Also, there is no editor in the default image, so I added a couple.

Also added some notes to the readme to talk about users and passwords and who was able to ssh into the image.